chore: move gitea-client to api service
