+++ .dockerignore
@@ -3,8 +3,6 @@
 !bun.lock
 !Dockerfile
 !packages/
-!packages/gitea-client/
-!packages/gitea-client/**
 !services/
 !services/api/
 !services/api/**
