+++ .github/workflows/deploy.yml
@@ -6,7 +6,6 @@ on:
       - main
     paths:
       - "services/api/**"
-      - "packages/gitea-client/**"
       - "packages/api-schema/**"
     tags:
       - "v[0-9]*"
