+++ .gitignore
@@ -37,8 +37,8 @@ report.[0-9]_.[0-9]_.[0-9]_.[0-9]_.json
 
 # Generated files
 lighthouse.json
-packages/gitea-client/spec/openapi3.json
-packages/gitea-client/spec/swagger2.json
+services/api/gitea-client/spec/openapi3.json
+services/api/gitea-client/spec/swagger2.json
 conf/app.ini
 conf/app.prod.ini
