+++ packages/editor/Editor.tsx
@@ -93,7 +93,7 @@ import { VersionHistory } from "./extensions/VersionHistory";
 import { Conflict } from "./extensions/conflict";
 import { gitService } from "./services/GitService";
 import { sanitizeHtml, sanitizeProseMirrorJson } from "../utils/sanitizer";
-import type { ApprovalState } from "../gitea-client/pullRequests";
+import type { ApprovalState } from "../../services/api/gitea-client/pullRequests";
 
 // --- Types ---
 interface ToolbarProps {
