+++ packages/editor/extensions/ApprovalStatus/ApprovalStatusBanner.tsx
@@ -1,4 +1,4 @@
-import type { ApprovalState } from "../../../gitea-client/pullRequests";
+import type { ApprovalState } from "../../../../services/api/gitea-client/pullRequests";
 
 export interface ApprovalStatusBannerProps {
   approvalState: ApprovalState;
