+++ services/api/Dockerfile
@@ -1,7 +1,6 @@
 FROM oven/bun:1 AS builder
 WORKDIR /app
 COPY package.json bun.lock ./
-COPY packages/gitea-client ./packages/gitea-client
 COPY services/api ./services/api
 RUN bun install --frozen-lockfile --production
