+++ services/api/gitea-client/spec/extensions.json
