+++ tests/document-download.pw.ts
@@ -23,7 +23,7 @@ import {
   getPullRequestForBranch,
   listPullRequests,
   submitReview,
-} from "../packages/gitea-client/pullRequests";
+} from "../services/api/gitea-client/pullRequests";
 import {
   createBobClient,
   expectedPrefilledDocumentName,
